requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cryptography==44.0.2
cssselect2==0.8.0
Django==5.1.7
django-cleanup==9.0.0
django-cotton==2.0.1
django-filter==25.1
django-solo==2.4.0
fonttools==4.57.0
pillow==11.1.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.0
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
